moves scripts to cmd folder
